CLI: Removed BackupCommand and moved commands in commands.php
