+++ cli/commands.php
@@ -5,47 +5,70 @@ namespace Unyson\Extension\Backups;
 
 class Command extends \Unyson\Extension\Command {
 
-	public function __construct( $name = 'backups' ) {
-		parent::__construct( $name );
-
-		spl_autoload_register( function ( $class ) {
-			switch ( $class ) {
-				case 'Unyson\Extension\Backups\BackupCommand' :
-					include_once dirname( __FILE__ ) . '/BackupCommand.php';
-					break;
-				case 'Unyson\Extension\Backups\DemoCommand' :
-					include_once dirname( __FILE__ ) . '/DemoCommand.php';
-					break;
-			}
-		} );
-	}
-
-	public function backup( $args, $options = array() ) {
-		$this->get_backup()->run_command( array_shift( $args ), $args, $options );
-	}
-
-	public function demo( $args, $options = array() ) {
-		$this->get_demo()->run_command( array_shift( $args ), $args, $options );
-	}
-
 	/**
-	 * @return BackupCommand
+	 * @param array $_
+	 * @param array $options
+	 *
+	 * @subcommand list
 	 */
-	protected function get_backup() {
-		return new BackupCommand();
+	public function _list( $_, $options = array() ) {
+		$archives = array_map(
+			array( $this, 'get_backup_file_data' ),
+			$this->get_backups()
+		);
+
+		\WP_CLI\Utils\format_items(
+			'table',
+			$archives,
+			array( 'Name', 'Time', 'Type' )
+		);
 	}
 
-	/**
-	 * @return DemoCommand
-	 */
-	protected function get_demo() {
-		return new DemoCommand();
+	public function create( $args, $options = array() ) {
+		$this
+			->get_ext()
+			->tasks()
+			->do_backup( isset( $options['full'] ) );
+
+		$this->message( "Backup successfully created" );
 	}
 
-	/**
-	 * @return \FW_Extension_Backups
-	 */
-	protected function get_ext() {
-		return fw_ext( 'backups' );
+	public function restore( $args, $options = array() ) {
+		$id = array_shift( $args );
+		try {
+			$backup = $this->get( $id );
+		} catch ( \Exception $e ) {
+			$this->error( "Backup $id doesn't seem to exist" );
+		}
+
+		$this->initialize_fs();
+		$this
+			->get_ext()
+			->tasks()
+			->do_restore( $backup['full'], $backup['path'] );
+
+		$this->message( "Backup $id was successfully restored" );
+	}
+
+	protected function get_backup_file_data( array $backup ) {
+		return array(
+			'Name' => basename( $backup['path'], '.zip' ),
+			'Type' => $backup['full'] ? 'Full' : 'Content',
+			'Time' => date( 'd M Y, H:i', $backup['time'] ),
+		);
+	}
+
+	protected function get_backups() {
+		return $this->get_ext()->get_archives();
+	}
+
+	protected function get( $id ) {
+		$backups = $this->get_backups();
+
+		if ( isset( $backups[ $id . '.zip' ] ) ) {
+			return $backups[ $id . '.zip' ];
+		}
+
+		throw new \Exception( "Backup $id cannot be found" );
 	}
 }
